Open storage editor by clicking the storage on the map
The per-storage editor used to be reachable only through the global
"Configured storages" list in settings, which forced the user to
hunt for the right tile coordinates. A new click handler now opens
that same editor for whatever storage is under the cursor when the
player taps Binding.select.

To stay out of the way:
- the handler skips the click while the player is placing a block
  (Vars.control.input.block != null) or while the cursor is over any
  scene actor (Core.scene.hit), so dialogs and HUD widgets keep
  consuming their own clicks first
- it ignores anything that isn't a player-team managed StorageBlock,
  so cores, enemy crates and non-storage tiles are untouched
- a new "eui-storage-click-ui" checkbox (default on) lets the player
  turn the click hijack off entirely

The editor itself was extracted from settings-ui.js into a shared
ui/dialogs/storage-edit-dialog module so both the click handler and
the storages list reuse the same code, and the editor now also
shows the live in-storage stock next to each item field.

Bumped to v0.92, translations updated (en / ru / uk_UA).
