+++ bundles/bundle.properties
@@ -30,6 +30,7 @@ setting.eui-auto-collect-factory.name = Auto-collect from factories
 setting.eui-auto-collect-drill.name = Auto-collect from drills
 setting.eui-collect-threshold.name = Auto-collect pickup threshold (% of producer capacity)
 setting.eui-storage-fill.name = Keep configured storages stocked
+setting.eui-storage-click-ui.name = Click a storage to open its config
 setting.eui-auto-pilot.name = Auto-pilot delivery (drone routes producer -> consumer by itself)
 setting.eui-DragBlock.name = Drag core to create vault
 setting.eui-DragPathfind.name = Drag conveyor to start pathfinding
