+++ bundles/bundle_ru.properties
@@ -30,6 +30,7 @@ setting.eui-auto-collect-factory.name = Автозабор с заводов
 setting.eui-auto-collect-drill.name = Автозабор с буров
 setting.eui-collect-threshold.name = Порог автозабора (% от ёмкости производителя)
 setting.eui-storage-fill.name = Поддерживать запас в настроенных хранилищах
+setting.eui-storage-click-ui.name = Открывать настройки хранилища по клику
 setting.eui-auto-pilot.name = Автопилот доставки (дрон сам летит производитель -> потребитель)
 setting.eui-DragBlock.name = Создавать хранилища при перетаскивании ядра
 setting.eui-DragPathfind.name = Использовать поиск пути при перетаскивании конвейеров
