+++ bundles/bundle_uk_UA.properties
@@ -30,6 +30,7 @@ setting.eui-auto-collect-factory.name = Автозабір із заводів
 setting.eui-auto-collect-drill.name = Автозабір із бурів
 setting.eui-collect-threshold.name = Поріг автозбору (% від місткості виробника)
 setting.eui-storage-fill.name = Підтримувати запас у налаштованих сховищах
+setting.eui-storage-click-ui.name = Відкривати налаштування сховища кліком
 setting.eui-auto-pilot.name = Автопілот доставки (дрон сам летить виробник -> споживач)
 setting.eui-DragBlock.name = Будувати склади при перетягуванні ядра
 setting.eui-DragPathfind.name = Перетягніть конвеєр, щоби почати пошук шляху
