+++ scripts/ui/other/settings-ui.js
@@ -2,6 +2,7 @@ const coreLimits = require("extended-ui/interact/core-limits");
 const collectConfig = require("extended-ui/interact/collect-config");
 const storageConfig = require("extended-ui/interact/storage-config");
 const storageFill = require("extended-ui/interact/storage-fill");
+const storageEditDialog = require("extended-ui/ui/dialogs/storage-edit-dialog");
 const iconsUtil = require("extended-ui/utils/icons");
 
 Events.on(EventType.ClientLoadEvent, () => {
@@ -54,6 +55,7 @@ Events.on(EventType.ClientLoadEvent, () => {
         contentTable.checkPref("eui-auto-collect-drill", false);
         contentTable.sliderPref("eui-collect-threshold", 50, 0, 100, 5, i => i + " %");
         contentTable.checkPref("eui-storage-fill", false);
+        contentTable.checkPref("eui-storage-click-ui", true);
         contentTable.checkPref("eui-auto-pilot", false);
         contentTable.sliderPref("eui-action-delay", 500, 0, 3000, 25, i => i + " ms");
         if (!Vars.mobile) {
@@ -263,72 +265,12 @@ function buildStorageListDialog() {
             : Core.bundle.get("eui.storage.row-empty")).pad(4);
 
         parent.button(Icon.pencil, Styles.cleari, () => {
-            buildStorageEditDialog(building, () => rebuild()).show();
+            storageEditDialog.build(building, () => rebuild()).show();
         }).size(36).pad(4);
 
         parent.row();
     }
 
     dialog.shown(() => rebuild());
-    return dialog;
-}
-
-function buildStorageEditDialog(building, onClose) {
-    const block = building.block;
-    const dialog = new BaseDialog(block.localizedName + " (" + building.tile.x + ", " + building.tile.y + ")");
-    dialog.addCloseButton();
-
-    dialog.buttons.button(Core.bundle.get("eui.storage.clear-all"), () => {
-        Vars.ui.showConfirm(
-            Core.bundle.get("eui.storage.clear-all"),
-            Core.bundle.get("eui.storage.clear-confirm"),
-            () => {
-                Vars.content.items().each(item => storageConfig.setThreshold(building, item, 0));
-                rebuild();
-            }
-        );
-    }).size(240, 60);
-
-    let listTable = null;
-    dialog.cont.add(Core.bundle.get("eui.storage.edit-hint")).width(580).wrap().pad(6).get().setAlignment(Align.center);
-    dialog.cont.row();
-    dialog.cont.pane(t => { listTable = t; t.top(); }).grow().maxHeight(540);
-
-    function rebuild() {
-        if (!listTable) return;
-        listTable.clearChildren();
-        const items = [];
-        Vars.content.items().each(item => items.push(item));
-        items.sort((a, b) => a.id - b.id);
-        for (let i = 0; i < items.length; i++) {
-            addItemRow(listTable, items[i]);
-        }
-    }
-
-    function addItemRow(parent, item) {
-        parent.image(iconsUtil.getByName(item.name)).size(32).pad(4);
-        parent.add(item.localizedName).left().width(140).pad(4);
-
-        const fieldCell = parent.field(storageConfig.getThreshold(building, item) + "", text => {
-            const v = parseInt(text);
-            if (!isNaN(v)) {
-                const clamped = Math.max(0, Math.min(storageConfig.MAX_THRESHOLD, v));
-                storageConfig.setThreshold(building, item, clamped);
-            }
-        });
-        fieldCell.valid(text => /^\d+$/.test(text) && parseInt(text) <= storageConfig.MAX_THRESHOLD);
-        fieldCell.width(110).pad(4);
-        const fieldElement = fieldCell.get();
-
-        parent.button(Icon.cancel, Styles.cleari, () => {
-            storageConfig.setThreshold(building, item, 0);
-            fieldElement.setText("0");
-        }).size(36).pad(4);
-
-        parent.row();
-    }
-
-    dialog.shown(() => rebuild());
-    if (onClose) dialog.hidden(() => onClose());
     return dialog;
 }
